Added implementations for facebook objects

+++ FacebookWinFormsApp/IFacebookObjectAdapter.cs
@@ -10,6 +10,6 @@ namespace BasicFacebookFeatures
     {
         String Text { get; }
         String ImageUrl { get; }
-        int LikesCount { get; }
+        long LikesCount { get; }
     }
 }
